src/python/requirements.txt
pyzmq==14.7.0

